+++ src/App.js
@@ -1,9 +1,9 @@
 import "./App.css";
 import { getLocal, setLocal } from "./Helpers/services";
-import Header from "./Header";
-import TransactionGenerator from "./TransactionsGenerator";
-import TransactionList from "./Transactions";
-import sampleTransactions from "./data.json";
+import Header from "./Components/Header";
+import TransactionGenerator from "./Components/TransactionsGenerator";
+import TransactionList from "./Components/Transactions";
+import sampleTransactions from "./Assets/JSON/data.json";
 import { useState, useEffect } from "react";
 
 export default function App() {
@@ -88,13 +88,14 @@ export default function App() {
           data={transactions}
           onTransactionAdd={handleTransactionAdd}
         />
+        <button onClick={() => setCount(count + 1)}>EXPAND</button>
+
         <TransactionList
           data={handleTransactionsDisplay(transactions, count)}
           dateFormatter={getFormattedDate}
           currencyFormatter={getFormattedAmount}
           transactionsGrouper={handleTransactionsGroup}
         />
-        <button onClick={() => setCount(count + 1)}>EXPAND</button>
       </main>
     </div>
   );
